Update the accepts_content_type to be more explicit
Bloqs does not just assume that any fieldtype returning true here is supported. Fieldtype developers should opt in to Bloqs support for a variety of reasons. The same goes for the new Fluid type. Adding a content type to the list shows that the fieldtype has been tested to support that content type.